requirements.txt
accelerate==0.25.0
aiofiles==23.2.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
arrow==1.3.0
attrs==23.1.0
backoff==2.2.1
blinker==1.4
boilerpy3==1.0.7
build==1.0.3
CacheControl==0.13.1
cattrs==23.2.3
certifi==2022.12.7
charset-normalizer==2.1.1
cleo==2.1.0
click==8.1.7
colorama==0.4.6
command-not-found==0.3
contourpy==1.2.0
crashtest==0.4.1
cryptography==3.4.8
cycler==0.12.1
dbus-python==1.2.18
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.1
docopt==0.6.2
dulwich==0.21.7
Events==0.5
exceptiongroup==1.2.0
faiss-cpu==1.7.2
farm-haystack==1.23.0
fastapi==0.108.0
fastjsonschema==2.19.0
ffmpy==0.3.1
filelock==3.13.1
fonttools==4.47.0
fsspec==2023.12.2
gradio_client==0.8.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httplib2==0.20.2
httpx==0.26.0
huggingface-hub==0.20.1
idna==3.4
importlib-metadata==7.0.1
importlib-resources==6.1.1
inflect==7.0.0
installer==0.7.0
jaraco.classes==3.3.0
jeepney==0.7.1
Jinja2==3.1.2
jinja2-time==0.2.0
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
keyring==24.3.0
kiwisolver==1.4.5
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-imports==0.3.1
make==0.1.6.post2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.0.7
netifaces==0.11.0
networkx==3.0
nltk==3.8.1
num2words==0.5.13
numpy==1.24.1
oauthlib==3.2.0
orjson==3.9.10
packaging==23.2
pandas==2.1.4
pexpect==4.9.0
Pillow==9.3.0
pkginfo==1.9.6
platformdirs==3.11.0
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
posthog==3.1.0
prompthub-py==4.0.0
protobuf==4.25.1
psutil==5.9.7
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pycairo==1.20.1
pydantic==1.10.13
pydantic_core==2.14.6
pydub==0.25.1
Pygments==2.17.2
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
pyproject_hooks==1.0.0
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
quantulum3==0.9.0
rank-bm25==0.2.2
rapidfuzz==3.5.2
referencing==0.32.0
regex==2023.12.25
requests==2.31.0
requests-cache==0.9.8
requests-toolbelt==1.0.0
rich==13.7.0
rpds-py==0.15.2
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
semantic-version==2.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.50
SQLAlchemy-Utils==0.41.1
sseclient-py==1.8.0
starlette==0.32.0.post1
sympy==1.12
systemd-python==234
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.5.2
tokenizers==0.15.0
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.0
torch==2.1.2+cu118
torchaudio==2.1.2+cu118
torchvision==0.16.2+cu118
tqdm==4.66.1
transformers==4.35.2
triton==2.1.0
trove-classifiers==2023.11.29
typer==0.9.0
types-python-dateutil==2.8.19.14
typing_extensions==4.9.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
url-normalize==1.4.3
urllib3==1.26.13
uvicorn==0.25.0
virtualenv==20.25.0
wadllib==1.3.6
websockets==11.0.3
xkit==0.0.0
zipp==1.0.0
